+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- `opentelemetry-instrumentation-requests` Added support for post-injection-hook.
+  ([#3657](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3657))
+
 ## Version 1.36.0/0.57b0 (2025-07-29)
 
 ### Fixed